Fixing the issue in struct removal and usage of struct in the type mapper.

+++ modules/web/js/ballerina/views/input-struct-view.js
@@ -15,8 +15,8 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-define(['lodash', 'log','./ballerina-view','./../ast/resource-parameter', 'typeMapper','constants'],
-    function (_, log, BallerinaView,ResourceParameter, TypeMapperRenderer,Constants) {
+define(['lodash', 'log', './ballerina-view', './../ast/resource-parameter', 'typeMapper', 'constants'],
+    function (_, log, BallerinaView, ResourceParameter, TypeMapperRenderer, Constants) {
 
         var InputStructView = function (args) {
             BallerinaView.call(this, args);
@@ -25,7 +25,7 @@ define(['lodash', 'log','./ballerina-view','./../ast/resource-parameter', 'typeM
             this._onDisconnectInstance = _.get(args, 'onDisconnectInstance', {});
             this._sourceInfo = _.get(args, 'sourceInfo', {});
 
-            if (_.isNil(this.getModel()) || !(this._model instanceof ResourceParameter)) {
+            if (!_.isNil(this.getModel()) && !(this._model instanceof ResourceParameter)) {
                 log.error("Resource parameter is undefined or is of different type." + this.getModel());
                 throw "Resource parameter is undefined or is of different type." + this.getModel();
             }
@@ -45,24 +45,30 @@ define(['lodash', 'log','./ballerina-view','./../ast/resource-parameter', 'typeM
          */
         InputStructView.prototype.render = function (diagramRenderingContext, mapper) {
             var self = this;
-            this._diagramRenderingContext = diagramRenderingContext;
-            var typeStructSchema = this.getSourceInfo().sourceStruct;
-            var typeStructName = this.getSourceInfo().sourceStructName;
-            var previousSelection = this.getSourceInfo()[TYPE_MAPPER_COMBOBOX_PREVIOUS_SELECTION];
-            var isAlreadtRenderedInTarget = this.getSourceInfo()[TYPE_MAPPER_COMBOBOX_SOURCE_IS_ALREADY_RENDERED_IN_TARGET];
-
-            if(!mapper) {
+            if (!mapper) {
                 mapper = new TypeMapperRenderer(self.getOnConnectInstance(), self.getOnDisconnectInstance(), this._parentView);
                 this._parentView._typeMapper = mapper;
             }
+            if (this.getSourceInfo().sourceStruct) {
+                this._diagramRenderingContext = diagramRenderingContext;
+                var typeStructSchema = this.getSourceInfo().sourceStruct;
+                var typeStructName = this.getSourceInfo().sourceStructName;
+                var previousSelection = this.getSourceInfo()[TYPE_MAPPER_COMBOBOX_PREVIOUS_SELECTION];
+                var isAlreadtRenderedInTarget = this.getSourceInfo()[TYPE_MAPPER_COMBOBOX_SOURCE_IS_ALREADY_RENDERED_IN_TARGET];
 
-            if(!_.isUndefined(previousSelection) && previousSelection != TYPE_MAPPER_COMBOBOX_DEFAULT_SELECTION){
+                if (!_.isUndefined(previousSelection) && previousSelection != TYPE_MAPPER_COMBOBOX_DEFAULT_SELECTION) {
+                    mapper.removeStruct(previousSelection);
+                }
+                if (!_.isUndefined(isAlreadtRenderedInTarget)) {
+                    mapper.removeStruct(typeStructName);
+                }
+                if (!_.isUndefined(typeStructSchema)) {
+                    mapper.addSourceStruct(typeStructSchema.getAttributesArray(), this.getModel());
+                }
+            } else {
+                var previousSelection = this.getSourceInfo()[TYPE_MAPPER_COMBOBOX_PREVIOUS_SELECTION];
                 mapper.removeStruct(previousSelection);
             }
-            if(!_.isUndefined(isAlreadtRenderedInTarget)){
-                mapper.removeStruct(typeStructName);
-            }
-            mapper.addSourceStruct(typeStructSchema.getAttributesArray(),this.getModel());
         };
 
         /**
@@ -116,4 +122,4 @@ define(['lodash', 'log','./ballerina-view','./../ast/resource-parameter', 'typeM
         };
 
         return InputStructView;
-});
+    });
